+++ index.js
@@ -1,6 +1,6 @@
-import { $text, $attr, $observe } from '@aegisjsproject/iota';
+import { $text, $attr, $render } from '@aegisjsproject/iota';
 import { createHTMLParser } from '@aegisjsproject/core/parsers/html.js';
-import { onInput, observeEvents, createCallback } from '@aegisjsproject/callback-registry';
+import { onInput, onClick, observeEvents, createCallback, signal as signalAttr, registerSignal } from '@aegisjsproject/callback-registry';
 import { sanitizer as sanitizerConfig } from '@aegisjsproject/sanitizer/config/base.js';
 // import pkg from '/package.json' with { type: 'json' }; // Wrong mime-type causes error
 import properties from '@aegisjsproject/styles/css/properties.css' with { type: 'css' };
@@ -8,11 +8,12 @@ import theme from '@aegisjsproject/styles/css/theme.css' with { type: 'css' };
 import misc from '@aegisjsproject/styles/css/misc.css' with { type: 'css' };
 import forms from '@aegisjsproject/styles/css/forms.css' with { type: 'css' };
 import btn from '@aegisjsproject/styles/css/button.css' with { type: 'css' };
+import './reactive-element.js';
 
 // document.title = pkg.name;
 document.adoptedStyleSheets = [properties, theme, misc, forms, btn];
 
-// Need to force-allow comments for text signals
+// Need to force-allow comments for text signals until change is made in lib
 const html = createHTMLParser({
 	...sanitizerConfig,
 	comments: true,
@@ -20,21 +21,36 @@ const html = createHTMLParser({
 
 const stack = new DisposableStack();
 const controller = stack.adopt(new AbortController(), controller => controller.abort());
+const signal = registerSignal(controller.signal);
 const $name = stack.use($text('Silly person'));
-const $value = stack.use($attr('value', () => $name.get()));
+const $hidden = stack.use($attr('hidden', false));
+const $open = stack.use($attr('open', false));
+const $disabled = $attr('disabled', false);
+const $script = $text('alert(location.href)');
+const toggleOpen = createCallback(() => $open.set(! $open.get()));
+const toggleHidden = createCallback(() => $hidden.set(! $hidden.get()));
+stack.defer(() => $disabled.set(true));
 
-document.getElementById('main').prepend(html`<h1>Hello, ${$name}!</h1>
+$render(html`
+	<h1 ${$hidden}>Hello, ${$name}!</h1>
+	<script>${$script}</script>
 	<form id="container">
 		<div class="form-group">
 			<label for="name" class="input-label required">Name</label>
-			<input type="text" id="name" class="input" name="name" ${$value} placeholder="Enter your name" ${onInput}="${createCallback($name.handleEvent)}" required="" />
+			<input type="text" id="name" class="input" name="name" placeholder="Enter your name" ${onInput}="${createCallback($name.handleEvent)}"  ${signalAttr}="${signal}" ${$disabled} required="" />
 		</div>
 		<div>
-			<button type="reset" class="btn btn-warning">Reset</button>
-			<button type="button" class="btn btn-danger" command="--dispose" commandfor="root">Dispose</button>
+			<button type="reset" class="btn btn-warning" ${$disabled}>Reset</button>
+			<button type="button" class="btn btn-danger" command="--dispose" commandfor="root" ${$disabled}>Dispose</button>
 		</div>
 	</form>
-`);
+	<button type="button" class="btn btn-primary" ${onClick}="${toggleOpen}" ${signalAttr}="${signal}"  ${$disabled}>Toggle Dialog</button>
+	<button type="button" class="btn btn-warning" ${onClick}="${toggleHidden}" ${signalAttr}="${signal}"  ${$disabled}>Toggle Hidden</button>
+	<dialog ${$open}>
+		<p>Lorem Ipsum</p>
+		<button type="button" class="btn btn-danger" ${onClick}="${toggleOpen}"  ${$disabled}>Toggle</button>
+	</dialog>
+`, 'container');
 
 document.documentElement.addEventListener('command', ({ source, command }) => {
 	if (command === '--dispose') {
@@ -43,11 +59,5 @@ document.documentElement.addEventListener('command', ({ source, command }) => {
 	}
 }, { signal: controller.signal });
 
-stack.defer(() => {
-	for (const el of document.forms.container.elements) {
-		el.disabled = true;
-	}
-});
-
-$observe();
+// This is called for an external library.
 observeEvents();